move util methods to separate file
