+++ sidechain_cli/utils/config_file.py
@@ -1,4 +1,4 @@
-"""Methods to do with the config file."""
+"""ConfigFile helper class."""
 
 from __future__ import annotations
 
@@ -61,19 +61,3 @@ class ConfigFile:
         """Write the ConfigFile data to file."""
         with open(_CONFIG_FILE, "w") as f:
             json.dump(self.to_dict(), f, indent=4)
-
-
-def _get_config() -> ConfigFile:
-    return ConfigFile.from_file()
-
-
-def add_chain(chain_data: Dict[str, Any]) -> None:
-    """
-    Add a chain's data to the config file.
-
-    Args:
-        chain_data: The data of the chain to add.
-    """
-    conf = _get_config()
-    conf.chains.append(chain_data)
-    conf.write_to_file()
